### 2022-09-24 fd0a7e7
:abc: adding missing DOIs found in openjournals/joss-reviews#4788
