+++ docs/references.bib
@@ -80,6 +80,7 @@
 	author = {Ide, Nancy M. and Sperberg-McQueen, C.M.},
 	year = {1995},
 	pages = {5--15},
+	doi = {10.1007/978-94-011-0325-1_2},
 }
 
 
@@ -179,6 +180,7 @@
 	year = {2015},
 	keywords = {Bible, Criticism, Textual, Criticism, interpretation, etc. History},
 	annote = {Thesis (Ph.D.) - Duke University, 2012},
+	doi = {10.1628/978-3-16-153324-2},
 }
 
 
