### 2022-09-24 6d848f4
:abc: adding JOSS status badge (openjournals/joss-reviews#4788)
