### 2022-09-26 #4788 @rbturnbull
Here are some names for possible JOSS reviewers who use phylogenetics in similar ways to what we are doing in this paper (and have Python as one of their preferred languages):

- SimonGreenhill
- tresoldi
- Anaphory

Here are some other potential reviewers who use phylogenetics and have Python as one of their preferred languages.
- marekborowiec
- rachelss
- tkchafin
- petaripenev
- jmcbroome
