+++ .github/workflows/beast2.yml
@@ -1,0 +1,36 @@
+name: beast2
+
+on: [push]
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.10"]
+    steps:
+    - uses: actions/checkout@v3
+    - name: Install poetry
+      run: pipx install poetry
+    - name: Install dependencies for Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v3
+      with:
+        python-version: ${{ matrix.python-version }}
+        cache: 'poetry'
+    - name: Install dependencies for Java
+      uses: actions/setup-java@v3
+      with:
+        distribution: 'temurin'
+        java-version: '17'
+        cache: 'maven'
+    - run: poetry install
+    - name: Testing
+      run: |
+        poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast2_example.nexus
+    - name: Install Phylogenetics Package
+      run: |
+        git clone https://github.com/CompEvol/beast2.git
+        cd beast2
+    - name: Phylogenetics Run
+      run: |
+        release/Linux/jrebin/beauti --help
