+++ .github/workflows/beast2.yml
@@ -22,7 +22,6 @@ jobs:
       with:
         distribution: 'temurin'
         java-version: '17'
-        cache: 'maven'
     - run: poetry install
     - name: Testing
       run: |
