+++ .github/workflows/beast2.yml
@@ -29,7 +29,7 @@ jobs:
     - name: Install Phylogenetics Package
       run: |
         git clone https://github.com/CompEvol/beast2.git
-        cd beast2
     - name: Phylogenetics Run
       run: |
-        release/Linux/jrebin/beauti --help
+        cd beast2
+        ./release/Linux/jrebin/beauti --help
