+++ .github/workflows/beast2.yml
@@ -34,4 +34,4 @@ jobs:
         ant compile-all
     - name: Phylogenetics Run
       run: |
-        ./beast2/release/Linux/jrebin/beauti --help
+        ls -l beast2
