+++ .github/workflows/beast2.yml
@@ -29,9 +29,9 @@ jobs:
     - name: Install Phylogenetics Package
       run: |
         git clone https://github.com/CompEvol/beast2.git
-        git clone https://github.com/CompEvol/BeastFX.git
         cd beast2
-        ant compile-all
+        docker build
     - name: Phylogenetics Run
       run: |
-        ls -l beast2/build/beast
+        cd beast2
+        docker run beast2-test test-beauti
