+++ .github/workflows/beast2.yml
@@ -26,12 +26,15 @@ jobs:
     - name: Testing
       run: |
         poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast2_example.nexus
-    - name: Install Phylogenetics Package
+    - name: Install Phylogenetics Packag
+      uses: dsaltares/fetch-gh-release-asset@master
+        with:
+          repo: 'CompEvol/beast2'
+          version: 'tags/v2.7.0'
+          file: 'BEAST.v2.7.0.Linux.x86.tgz'
+          token: ${{ secrets.GITHUB_TOKEN }}
       run: |
-        git clone https://github.com/CompEvol/beast2.git
-        cd beast2
-        docker build -t beast2
+        tar -zxvf BEAST.v2.7.0.Linux.x86.tgz
     - name: Phylogenetics Run
       run: |
-        cd beast2
-        docker run beast2-test
+        ls -l BEAST.v2.7.0.Linux.x86/bin
