+++ .github/workflows/beast2.yml
@@ -26,15 +26,14 @@ jobs:
     - name: Testing
       run: |
         poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast2_example.nexus
-    - name: Install Phylogenetics Packag
+    - name: Install Phylogenetics Package
       uses: dsaltares/fetch-gh-release-asset@master
         with:
           repo: 'CompEvol/beast2'
           version: 'tags/v2.7.0'
           file: 'BEAST.v2.7.0.Linux.x86.tgz'
           token: ${{ secrets.GITHUB_TOKEN }}
-      run: |
-        tar -zxvf BEAST.v2.7.0.Linux.x86.tgz
+    - run: tar -zxvf BEAST.v2.7.0.Linux.x86.tgz
     - name: Phylogenetics Run
       run: |
         ls -l BEAST.v2.7.0.Linux.x86/bin
