+++ .github/workflows/beast2.yml
@@ -28,11 +28,10 @@ jobs:
         poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast2_example.nexus
     - name: Install Phylogenetics Package
       uses: dsaltares/fetch-gh-release-asset@master
-        with:
-          repo: 'CompEvol/beast2'
-          version: 'tags/v2.7.0'
-          file: 'BEAST.v2.7.0.Linux.x86.tgz'
-          token: ${{ secrets.GITHUB_TOKEN }}
+      with:
+        repo: 'CompEvol/beast2'
+        version: 'tags/v2.7.0'
+        file: 'BEAST.v2.7.0.Linux.x86.tgz'
     - run: tar -zxvf BEAST.v2.7.0.Linux.x86.tgz
     - name: Phylogenetics Run
       run: |
