+++ .github/workflows/beast2.yml
@@ -35,4 +35,4 @@ jobs:
     - run: tar -zxvf BEAST.v2.7.0.Linux.x86.tgz
     - name: Phylogenetics Run
       run: |
-        ls -l beast/bin
+        beast/bin/beauti --help
