+++ .github/workflows/beast2.yml
@@ -35,4 +35,5 @@ jobs:
     - run: tar -zxvf BEAST.v2.7.0.Linux.x86.tgz
     - name: Phylogenetics Run
       run: |
+        ls -l beast/lib/
         java -cp beast/lib/launcher.jar beast.app.beauti.BeautiLauncher --help
