+++ .github/workflows/beast2.yml
@@ -35,5 +35,4 @@ jobs:
     - run: tar -zxvf BEAST.v2.7.0.Linux.x86.tgz
     - name: Phylogenetics Run
       run: |
-        ls -l beast/lib/
-        java -cp beast/lib/launcher.jar BEAST.app.jar --help
+        ls beast/bin
