+++ .github/workflows/beast.yml
@@ -25,11 +25,14 @@ jobs:
     - run: poetry install
     - name: Testing
       run: |
-        poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast_example.nexus
+        poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast2_example.nexus
     - name: Install Phylogenetics Package
-      run: |
-        git clone https://github.com/beast-dev/beast-mcmc
+      uses: dsaltares/fetch-gh-release-asset@master
+      with:
+        repo: 'beast-dev/beast-mcmc'
+        version: 'SARS-CoV-2-origins'
+        file: 'SARS-CoV-2-origins.tar.gz'
+    - run: tar -zxvf SARS-CoV-2-origins.tar.gz
     - name: Phylogenetics Run
       run: |
-        beast-mcmc/release_beastgen/common/bin/beastgen -chain_length=100000 beast_example.template beast_example.nexus beast_example.xml
-        tail beast_example.xml
+        ls -l
