+++ .github/workflows/beast.yml
@@ -30,7 +30,7 @@ jobs:
       uses: dsaltares/fetch-gh-release-asset@master
       with:
         repo: 'beast-dev/beast-mcmc'
-        version: 'SARS-CoV-2-origins'
+        version: 'tags/SARS-CoV-2-origins'
         file: 'SARS-CoV-2-origins.tar.gz'
     - run: tar -zxvf SARS-CoV-2-origins.tar.gz
     - name: Phylogenetics Run
