+++ .github/workflows/beast.yml
@@ -30,9 +30,9 @@ jobs:
       uses: dsaltares/fetch-gh-release-asset@master
       with:
         repo: 'beast-dev/beast-mcmc'
-        version: 'tags/SARS-CoV-2-origins'
-        file: 'SARS-CoV-2-origins.tar.gz'
-    - run: tar -zxvf SARS-CoV-2-origins.tar.gz
+        version: 'tags/sars-cov-2-origins'
+        file: 'sars-cov-2-origins.tar.gz'
+    - run: tar -zxvf sars-cov-2-origins.tar.gz
     - name: Phylogenetics Run
       run: |
         ls -l
