+++ .github/workflows/beast.yml
@@ -30,9 +30,9 @@ jobs:
       uses: dsaltares/fetch-gh-release-asset@master
       with:
         repo: 'beast-dev/beast-mcmc'
-        version: 'tags/sars-cov-2-origins'
-        file: 'sars-cov-2-origins.tar.gz'
-    - run: tar -zxvf sars-cov-2-origins.tar.gz
+        version: 'tags/v1.10.4'
+        file: 'BEASTv1.10.4.tgz'
+    - run: tar -zxvf BEASTv1.10.4.tgz
     - name: Phylogenetics Run
       run: |
         ls -l
