+++ .github/workflows/beast.yml
@@ -35,4 +35,4 @@ jobs:
     - run: tar -zxvf BEASTv1.10.4.tgz
     - name: Phylogenetics Run
       run: |
-        ls -l
+        ls -l BEASTv1.10.4/bin 
