+++ .github/workflows/beast.yml
@@ -27,12 +27,10 @@ jobs:
       run: |
         poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast2_example.nexus
     - name: Install Phylogenetics Package
-      uses: dsaltares/fetch-gh-release-asset@master
-      with:
-        repo: 'beast-dev/beast-mcmc'
-        version: 'tags/v1.10.4'
-        file: 'BEASTv1.10.4.tgz'
-    - run: tar -zxvf BEASTv1.10.4.tgz
+      run: |
+        git clone https://github.com/beast-dev/beast-mcmc.git
+        cd beast-mcmc
+        ant -pre-compile
     - name: Phylogenetics Run
       run: |
-        ls -l BEASTv1.10.4/bin 
+        ls -l beast-mcmc
