+++ .github/workflows/beast.yml
@@ -25,7 +25,7 @@ jobs:
     - run: poetry install
     - name: Testing
       run: |
-        poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast2_example.nexus
+        poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast_example.nexus
     - name: Install Phylogenetics Package
       run: |
         git clone https://github.com/beast-dev/beast-mcmc.git
@@ -33,5 +33,6 @@ jobs:
         ant
     - name: Phylogenetics Run
       run: |
-        cd beast-mcmc
-        release_beastgen/common/bin/beastgen -chain_length 100000 release_beastgen/common/templates/beast_example.template ../beast_example.nexus ../beast_example.xml
+        cp beast_example.nexus beast-mcmc/release_beastgen/common/bin
+        cd beast-mcmc/release_beastgen/common/bin
+        beastgen -chain_length 100000 beast_example.template beast_example.nexus beast_example.xml
