+++ .github/workflows/beast.yml
@@ -28,11 +28,11 @@ jobs:
         poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast_example.nexus
     - name: Install Phylogenetics Package
       run: |
-        git clone https://github.com/beast-dev/beast-mcmc.git
-        cd beast-mcmc
-        ant
+        git clone https://github.com/CompEvol/beast2.git
+        cd beast2
+        ant compile-all
     - name: Phylogenetics Run
       run: |
-        cp beast_example.nexus beast-mcmc/release_beastgen/common/bin
-        cd beast-mcmc/release_beastgen/common/bin
-        ./beastgen -chain_length 100000 beast_example.template beast_example.nexus beast_example.xml
+        ls -l beast2
+        ls -l beast2/build
+        java -cp /path/to/beast.jar beast.app.beauti.Beauti -template your_template.xml -nex your_nexus.nex -out out.xml -exitaction writexml
