+++ .github/workflows/beast.yml
@@ -34,5 +34,5 @@ jobs:
         ant compile-all
     - name: Phylogenetics Run
       run: |
-        find -i "beast.jar" -r .
+        find "beast.jar" -r .
         java -cp /path/to/beast.jar beast.app.beauti.Beauti -template your_template.xml -nex your_nexus.nex -out out.xml -exitaction writexml
