+++ .github/workflows/beast.yml
@@ -1,38 +1,0 @@
-name: beast
-
-on: [push]
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.10"]
-    steps:
-    - uses: actions/checkout@v3
-    - name: Install poetry
-      run: pipx install poetry
-    - name: Install dependencies for Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v3
-      with:
-        python-version: ${{ matrix.python-version }}
-        cache: 'poetry'
-    - name: Install dependencies for Java
-      uses: actions/setup-java@v3
-      with:
-        distribution: 'temurin'
-        java-version: '17'
-    - run: poetry install
-    - name: Testing
-      run: |
-        poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --no-labels --states-present --calibrate-dates example/ubs_ephesians.xml beast_example.nexus
-    - name: Install Phylogenetics Package
-      run: |
-        git clone https://github.com/CompEvol/BeastFX.git
-        git clone https://github.com/CompEvol/beast2.git
-        cd beast2
-        ant compile-all
-    - name: Phylogenetics Run
-      run: |
-        find . -name "beast.jar"
-        java -cp /path/to/beast.jar beast.app.beauti.Beauti -template your_template.xml -nex your_nexus.nex -out out.xml -exitaction writexml
