### 2022-10-10 #4788 @majensen
@tresoldi would you be willing to review [this work](#4788) for [JOSS](https://joss.theoj.org)? Your consideration is much appreciated! --@majensen
