### 2022-10-22 #4788 @majensen
@SimonGreenhill, would you have time to review [this work](#4788) for [JOSS](https://joss.theoj.org)? I would appreciate your expertise in using phylogenetics in linguistics.
Thanks for your consideration! --@majensen
