### 2022-10-25 #4879 @editorialbot
[REVIEW]: teiphy: A Python Package for Converting TEI XML Collations to NEXUS and Other Formats
**Submitting author:** @jjmccollum (<a href="http://orcid.org/0000-0002-5647-0365">Joey McCollum</a>)
**Repository:** https://github.com/jjmccollum/teiphy
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
