+++ pyproject.toml
@@ -28,7 +28,7 @@ python-slugify = "^6.1.2"
 
 [tool.poetry.dev-dependencies]
 Sphinx = "^5.1.1"
-pytest = "^7.1.2"
+pytest = "^7.2.0"
 coverage = "^6.4.4"
 nbsphinx = "^0.8.7"
 sphinx-rtd-theme = "^1.0.0"
