### 2022-10-28 #4879 @jjmccollum
@SimonGreenhill Thanks for your thorough and insightful comments! I'll open a branch or two for revisions to the paper and additions to the code. I don't think it should be too hard to add support for phy and fasta, and it would definitely be within the scope of the project to do so.
