### 2022-10-28 jjmccollum/teiphy#47 @rbturnbull
I think we should be able to do it with a callback in typer and using importlib_metadata. There is a way to do it with a standard python module but I think that's only from a recent python release. Have a look here:
https://github.com/rbturnbull/ausdex/blob/main/ausdex/main.py
