### 2022-10-28 jjmccollum/teiphy#47 @rbturnbull
also here: https://github.com/rbturnbull/torchapp/blob/master/torchapp/apps.py
