### 2022-10-28 jjmccollum/teiphy#48 @jjmccollum
Add support for PHYLIP and FASTA formats
Per the latest reviewer comments (https://github.com/openjournals/joss-reviews/issues/4879), support for PHYLIP and FASTA formats (used by programs like PHYLIP, RAxML, and IQTREE) would be convenient. Format parsing and output methods for these formats should be added to the code.

These changes can be made on the `code-revision` branch.
