### 2022-10-28 jjmccollum/teiphy#47 @jjmccollum
Thanks! I tried adapting the code in the first example, as it seemed cleaner. I added `from importlib.metadata import version` to the preamble. Then I added the following function:

```python
def version_callback(value: bool):
    if value:
       teiphy_version = version("teiphy")
        typer.echo(teiphy_version)
        raise typer.Exit()
```

And I added the following option to the app:

```python
version: bool = typer.Option(
    False,
    callback=version_callback,
    is_eager=True,
    help="Print the current version.",
),
```

Entering the command `poetry run teiphy --version` works as expected, but the program keeps printing the version number "0.1.0" to the console. The version number in `pyproject.toml` is "0.1.3", so now I'm trying to figure out what is causing the discrepancy. I cleaned out the `dist` folder and ran `poetry build` just to be safe, but I'm still getting a version number of "0.1.0".
