### 2022-10-28 8c7fd4f
Added more classifiers to pyproject.toml; added version checking to main.py, but need to debug (#47)
