### 2022-10-28 jjmccollum/teiphy#47 @rbturnbull
I'll check it out. By the way - what version of poetry are you using? I had a similar problem to this in another project before upgrading to poetry 2
