### 2022-10-28 jjmccollum/teiphy#47 @jjmccollum
Ah, perhaps that's the issue; I'm using version 1.1.13. I can try upgrading now.

If poetry 2 is necessary, is there a line in the `pyproject.toml` file that we should change accordingly? The following lines look like they might fit:

```
[build-system]
requires = ["poetry-core>=1.0.0"]
```
