### 2022-10-28 jjmccollum/teiphy#47 @rbturnbull
I think it's just that you need to clear out the old distribution files before the right version number is picked up. I don't think we need to require poetry 2 in pyproject.toml. Up to you...
