### 2022-10-28 jjmccollum/teiphy#47 @jjmccollum
Is there anything I need to clear out apart from the files in the `dist` directory? I've done that and run `poetry build`, but I still get the wrong version number.
