### 2022-10-28 jjmccollum/teiphy#47 @rbturnbull
it's working for me on the code-revision branch.

Try clearing out these directories in your virtual env: `.venv/lib/python3.9/site-packages/teiphy-*`
(wherever poetry installs your venv)

Then try `poetry install `
