### 2022-10-28 jjmccollum/teiphy#47 @jjmccollum
That did it! Thanks for suggesting clearing out the cache for the virtual environment.

Since we've established that this feature works, I'll go ahead and close this issue. I'm almost done implementing output methods for the PHYLIP and FASTA formats, but I'll have to add unit tests to ensure 100% coverage, and if it's easy to enough to add testing workflows for PHYLIP and/or RAxML, then we may want to do that. Once all of that is in place, we can merge `code-revision` into `main`.
