+++ teiphy/format.py
@@ -8,6 +8,8 @@ class FormatUnknownException(Exception):
 class Format(Enum):
     NEXUS = 'NEXUS'
     HENNIG86 = 'HENNIG86'
+    PHYLIP = ('PHYLIP',)
+    FASTA = ('FASTA',)
     CSV = 'CSV'
     TSV = 'TSV'
     EXCEL = 'EXCEL'
@@ -19,6 +21,10 @@ class Format(Enum):
             ".nex": cls.NEXUS,
             ".nexus": cls.NEXUS,
             ".nxs": cls.NEXUS,
+            ".ph": cls.PHYLIP,
+            ".phy": cls.PHYLIP,
+            ".fa": cls.FASTA,
+            ".fasta": cls.FASTA,
             ".tnt": cls.HENNIG86,
             ".csv": cls.CSV,
             ".tsv": cls.TSV,
