+++ tests/test_format.py
@@ -8,6 +8,10 @@ class FormatTestCase(unittest.TestCase):
         self.assertEqual(Format.infer(".nex"), Format.NEXUS)
         self.assertEqual(Format.infer(".nxs"), Format.NEXUS)
         self.assertEqual(Format.infer(".nexus"), Format.NEXUS)
+        self.assertEqual(Format.infer(".ph"), Format.PHYLIP)
+        self.assertEqual(Format.infer(".phy"), Format.PHYLIP)
+        self.assertEqual(Format.infer(".fa"), Format.FASTA)
+        self.assertEqual(Format.infer(".fasta"), Format.FASTA)
         self.assertEqual(Format.infer(".tnt"), Format.HENNIG86)
         self.assertEqual(Format.infer(".csv"), Format.CSV)
         self.assertEqual(Format.infer(".tsv"), Format.TSV)
