### 2022-10-28 f5b0694
:bug: Fixed malformatted lines in format.py
