### 2022-10-28 jjmccollum/teiphy#48 @jjmccollum
Okay, I've added the code and unit tests (and fixed an error I made that was causing that last testing workflow to fail). If it's not too hard, it would be nice to add testing workflows with RAxML (https://github.com/stamatak/standard-RAxML), which can parse both PHYLIP and FASTA inputs, to ensure that the alignments we generate are valid. @rbturnbull, do you think it would be possible for us to clone the RAxML repo, do a build, and run RAxML with both types of input in a test runner?
