### 2022-10-28 jjmccollum/teiphy#48 @rbturnbull
we can try it and build it and cache the result in github actions. Alternatively we can install it with conda: https://anaconda.org/bioconda/raxml
