### 2022-10-28 jjmccollum/teiphy#48 @jjmccollum
Okay, I got a RAxML workflow that tests both PHYLIP and FASTA formats working on `code-revision`! I will now merge that branch into `main`.
