### 2022-10-28 jjmccollum/teiphy#50 @jjmccollum
Update documentation
Per the reviewer comments at https://github.com/openjournals/joss-reviews/issues/4879, we should add documentation on how to run the unit tests. From the GitHub repository, this is done via the command

```bash
poetry run pytest
```

@rbturnbull, is this all we need to add to the documentation, or is there some other way that we can run the unit tests if we've installed `teiphy` via pip?
