### 2022-10-28 30b08bc
Updated readme and advanced usage documentation to reflect new features (#50)
