### 2022-10-29 jjmccollum/teiphy#52 @jjmccollum
Revise paper
Per the latest reviewer comments (https://github.com/openjournals/joss-reviews/issues/4879), we should expand the Use Case section to discuss the UBS dataset and the results depicted for it in Figure 1. We could include the results from other software in other figures to highlight the consistency of several groupings; including the output of `stemma` would also allow us to talk briefly about how known contamination in the dataset is accounted for).

Aside from this, several minor revisions were recommended. I have reproduced the checklist here:

- [x] I would like more detail on the example, preferably with worked through command line steps. This is already in the example section of the docs, so please add this into the paper too (especially the step of generating a nexus file, showing and explaining the command line arguments).

- [x] Figure 1, and discussion: can you please describe what this tree show us (e.g. do these groupings make sense, do they tell us something interesting about this document?)

- [x] L16: please define what the acronym "TEI XML" means.

- [x] L37 "since then, phylogenetic methods have quickly evolved (Felsenstein, 2004),". Please reword -- this reads as if phylogenetic methods were driven by textual analyses, whereas, these methods have been in continuous development since the 70s and 80s, and not for textual analyses.

- [x] L52 "great chasm .. fixed". Please edit - it's a bit over the top, and/or the verb "fixed" is awkward. Perhaps "exists" or "hinders the use of TEI XML in phylogenetic contexts".

- [x] L60: TNT's claims of 'remarkable performance' may have been true in 2008 but now there are plenty of genomic-scale methods out there that can outcompete TNT (something like RAXML-NG probably blows it out of water for example). A better argument for TNT is that it has a comprehensive of morphological state models that may be appropriate for digital humanities data. You later go on to use tools like IQ-Tree and MrBayes, so it'd be worth broadening this section to mentioning these tools up front too. Don't let people think your program is innately tied to "cladistic" tools and approaches (and I mean "cladistic" here in the pejorative sense).

- [x] L72: Might be worth noting that this provides an immediate pathway into all the machine learning tools available in python (e.g. scikit-learn, tensorflow etc).

- [x] L75-76: Can you briefly explain what 'lacunae' and 'retroversions' are (this also strengthens the case for your program). Another term 'corrector' is used later in the m.s. without explanation as well.

- [x] L79: what do you mean by 'soft decision'?
