### 2022-10-29 ba8815a
:abc: Did a pass at revising paper (#52)
