+++ paper.md
@@ -26,16 +26,15 @@ bibliography: docs/references.bib
 
 Textual scholars have been using phylogenetics to analyze manuscript traditions since the early 1990s [@roh_report_1992].
 Many standard phylogenetic software packages accept as input the NEXUS file format [@msm_nexus_1997]. 
-The `teiphy` program takes a collation of texts encoded using the Text Encoding Initiative (TEI) guidelines and converts it to a NEXUS file
-that can be used for phylogenetic analysis.
+The `teiphy` program takes a collation of texts encoded in TEI XML format and converts it to a NEXUS file (among other formats)
+amenable to phylogenetic analysis.
 The package can also convert to other formats such as Stephen C. Carlson’s [STEMMA](https://github.com/stemmatic/stemma) format or to a NumPy array [@numpy_2020].
 
 # Statement of Need
 
-The TEI aims to provide an international standard for digitally encoding textual information for the humanities [@ism_tei_1995]. 
-The TEI guidelines describe an XML format for encoding a critical apparatus [@tei_critical_apparatus].
-Due to its rich and well-documented set of elements for expressing a wide range of features in manuscript transcriptions, collations, and critical editions,
-TEI XML has become the _de facto_ format for textual data in the digital humanities [@fischer_representing_2020]. 
+For over a decade, the Text Encoding Initiative has endeavored to provide an international standard for digitally encoding textual information for the humanities [@ism_tei_1995]. 
+Their guidelines describe a standard format for encoding material details, textual transcriptions, and critical apparatuses [@tei_critical_apparatus].
+Due to its rich and well-documented set of elements for expressing a wide range of features in these settings, the Text Encoding Initiative’s Extensible Markup Language format (hereafter abbreviated TEI XML) has become the _de facto_ format for textual data in the digital humanities [@fischer_representing_2020]. 
 Its expressive power has proven increasingly valuable since its release, as scholars have learned—sometimes the hard way—that digital transcriptions and collations should
 
 1. preserve as much detail as they can from their material sources, including paratextual features; 
@@ -43,37 +42,40 @@ Its expressive power has proven increasingly valuable since its release, as scho
 3. describe uncertainties about a source’s contents as accurately as possible, allowing for degrees of uncertainty and multiple choices for disambiguations if necessary.
 
 These principles have much bearing on the editing of critical texts, a task fundamental to both digital humanities and classical philology.
-Within the digital humanities, phylogenetic algorithms have been popular approaches to this task.
+Within the digital humanities, phylogenetic algorithms developed in the context of evolutionary biology have been popular approaches to this task.
 Taking the most arduous part of reconstructing a textual tradition and delegating it to a computer proved to be a promising technique, and its successful demonstration with a portion of _The Canterbury Tales_ was a milestone in the development of the field [@bhbr_phylogeny_1998].
 Soon after this, the same methods were applied more comprehensively to the tradition of _Lanseloet van Denemerken_ in a work that would formalize many practical rules for computer-assisted textual criticism [@salemans_building_2000].
-Since then, phylogenetic methods have quickly evolved [@felsenstein_inferring_2004], and textual critics have adapted the improvements and even added their own innovations to make the process more suitable for their purposes [@swh_vorlage_2002; @swh_pathways_2004; @carlson_text_2015; @edmondson_analysis_2019; @turnbull_history_2020; @hyytiainen_acts_2021].
[…]
