+++ docs/references.bib
@@ -1,37 +1,37 @@
 @article{msm_nexus_1997,
-	author = {Maddison, David R. and Swofford, David L. and Maddison, Wayne P.},
-	title = {{NEXUS: An Extensible File Format for Systematic Information}},
-	journal = {Systematic Biology},
-	volume = {46},
-	number = {4},
-	pages = {590--621},
-	year = {1997},
-	month = {12},
-	abstract = "{NEXUS is a file format designed to contain systematic data for use by computer programs. The goals of the format are to allow future expansion, to include diverse kinds of information, to be independent of particular computer operating systems, and to be easily processed by a program. To this end, the format is modular, with a file consisting of separate blocks, each containing one particular kind of information, and consisting of standardized commands. Public blocks (those containing information utilized by several programs) house information about taxa, morphological and molecular characters, distances, genetic codes, assumptions, sets, trees, etc.; private blocks contain information of relevance to single programs. A detailed description of commands in public blocks is given. Guidelines are provided for reading and writing NEXUS files and for extending the format.}",
-	issn = {1063-5157},
-	doi = {10.1093/sysbio/46.4.590},
-	url = {https://doi.org/10.1093/sysbio/46.4.590},
-	eprint = {https://academic.oup.com/sysbio/article-pdf/46/4/590/19502018/46-4-590.pdf},
+    author = {Maddison, David R. and Swofford, David L. and Maddison, Wayne P.},
+    title = {{NEXUS: An Extensible File Format for Systematic Information}},
+    journal = {Systematic Biology},
+    volume = {46},
+    number = {4},
+    pages = {590--621},
+    year = {1997},
+    month = {12},
+    abstract = "{NEXUS is a file format designed to contain systematic data for use by computer programs. The goals of the format are to allow future expansion, to include diverse kinds of information, to be independent of particular computer operating systems, and to be easily processed by a program. To this end, the format is modular, with a file consisting of separate blocks, each containing one particular kind of information, and consisting of standardized commands. Public blocks (those containing information utilized by several programs) house information about taxa, morphological and molecular characters, distances, genetic codes, assumptions, sets, trees, etc.; private blocks contain information of relevance to single programs. A detailed description of commands in public blocks is given. Guidelines are provided for reading and writing NEXUS files and for extending the format.}",
+    issn = {1063-5157},
+    doi = {10.1093/sysbio/46.4.590},
+    url = {https://doi.org/10.1093/sysbio/46.4.590},
+    eprint = {https://academic.oup.com/sysbio/article-pdf/46/4/590/19502018/46-4-590.pdf},
 }
 
 @book{farris_hennig86_1988,
-	author = {Farris, James S.},
-	title = {{Hennig86, ver. 1.5. Program and Documentation}},
-	address = {Port Jefferson Station, {NY}},
-	publisher = {James S. Farris},
-	year = {1988},
+    author = {Farris, James S.},
+    title = {{Hennig86, ver. 1.5. Program and Documentation}},
+    address = {Port Jefferson Station, {NY}},
+    publisher = {James S. Farris},
+    year = {1988},
 }
 
 @article{gc_tnt_2016,
-	author = {Goloboff, Pablo A. and Catalano, Santiago A.},
-	title = {{TNT, Version 1.5, Including a Full Implementation of Phylogenetic Morphometrics}},
-	journal = {Cladistics},
-	volume = {32},
-	number = {3},
-	year = {2016},
-	pages = {221--238},
-	doi = {10.1111/cla.12160},
[…]
