### 2022-10-30 jjmccollum/teiphy#52 @jjmccollum
@rbturnbull I've pushed a first pass of revisions to the paper on the `paper-revision` branch. I don't remember where we accessed the generated PDF on GitHub, but if you do, any feedback or further changes from you would be appreciated! I added several citations, but some of the sources (like those from _Andrews University Seminary Studies_ and the _Journal of Machine Learning Research_)

If everything looks good, I'll make a pull request and close this issue.
