### 2022-10-30 jjmccollum/teiphy#50 @rbturnbull
Hi @jjmccollum - sorry about the delay. I think you can only run the tests when you clone the repo. The tests aren't included in the python package
