### 2022-10-30 jjmccollum/teiphy#52 @rbturnbull
Hi @jjmccollum - you can render the paper using the ``./paper.sh`` script in the base of the repo if you have docker running. Otherwise the CI pipeline for the docs renders and deploys the paper to https://jjmccollum.github.io/teiphy/joss-draft.pdf for every push to the ``main`` branch. Can either of those options work for you?
