### 2022-10-30 jjmccollum/teiphy#50 @jjmccollum
@rbturnbull Thanks, I just wanted to clarify that point. In that case, what I've added to the readme should be sufficient. I'll close this issue now.
