### 2022-10-30 jjmccollum/teiphy#52 @jjmccollum
Sure! I'll do a pull request to merge `paper-revision` with `main` so we can check the main paper on the `main` branch, but I'll leave this issue open until I have your feedback.
