### 2022-10-30 28989fb
:abc: Fixed asterisk after PAUP in paper (#52)
