### 2022-10-30 6d5f84e
:abc: Fixed some wording and formatting in paper (#52)
