+++ paper.md
@@ -71,7 +71,7 @@ Given the prevalence of efforts like these, the need for a means of converting T
 
 While the conversion process is a straightforward one for most collation data, various sources of ambiguity can make a one-to-one mapping of witnesses to readings impossible.
 One such source of ambiguity is lacunae, or gaps in the text due to erasure, faded ink, or damage to the page.
-Another is retroversions, or reconstructed readings in the original language of the text resulting from the back-translation subsequent translations of the text into other languages.
+Another is retroversions, or readings in the original language of the text reconstructed through the back-translation of subsequent versions of the text in other languages.
 Mechanisms for modeling ambiguous states resulting from situations like these exist in both TEI XML and NEXUS, and in both parsimony- and likelihood-based phylogenetic methods, ambiguities about the states at the leaves and even at the root of the tree can be encoded and leveraged in the inference process.
 For these reasons, it is imperative to ensure that these types of judgments, as well as other rich features from TEI XML, can be respected (and, where necessary, preserved) in the conversion process.
 
@@ -127,7 +127,7 @@ If, in variation units where manuscripts are corrected or have alternate reading
 (Note that because the `*` character is reserved on the command-line, we must place it between quotation marks directly after the `-s` flag.)
 
 When corrections are made to a manuscript, they are typically sporadic, and as a result, the text of corrector witnesses like 06C1 and 06C2 will tend to be too fragmentary to be useful for analysis.
-But if we wish to assume that each corrector tacitly adopted of all the readings from the previous hand that he or she did not change, then `teiphy` can “fill out” each corrector’s text using the text of the first hand (in the case of the first corrector) or the filled-out text of the previous corrector (for all subsequent correctors).
+But if we wish to assume that each corrector tacitly adopted all of the readings from the previous hand that he or she did not change, then `teiphy` can “fill out” each corrector’s text using the text of the first hand (in the case of the first corrector) or the filled-out text of the previous corrector (for all subsequent correctors).
 Thus, 06C1 would replicate the text of 06* (i.e., the first hand responsible for the text of 06) where it does not introduce its own readings, and 06C2 would then replicate the text of 06C1 where it does not introduce its own readings.
 If we want to apply this transformation during the conversion process, then we can specify this with the `--fill-correctors` flag.
 
@@ -137,7 +137,8 @@ Finally, we must specify the required arguments to `teiphy`, which are the input
 Note that we do not have to specify the desired output format explicitly; `teiphy` will determine from the output filename that it should write a NEXUS file.
 Combining the previous options and arguments, the complete command is
 ```
-teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --states-present example/ubs_ephesians.xml ubs_ephesians-iqtree.nexus
+teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac
+-s"*" -s T --fill-correctors --states-present example/ubs_ephesians.xml ubs_ephesians-iqtree.nexus
 ```
 
 If we pass the resulting NEXUS file to IQ-TREE and specify appropriate settings for our textual data (in this case, the Lewis Mk substitution model with ascertainment bias correction), we will get an output tree like the one shown in Figure 1.
@@ -153,7 +154,8 @@ Second, since multiple files are written for STEMMA input (namely, a collation f
 
 Combining the options and arguments we used before with these changes, the complete command is
 ```
-teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --format stemma example/ubs_ephesians.xml stemma_example
[…]
