### 2022-10-31 jjmccollum/teiphy#52 @jjmccollum
@rbturnbull Sorry for the additional changes on my end; once I could see the PDF, I made some further commits to fix wording and formatting problems I noticed in it. I'm done with edits, so feel free to take a look when you get a chance!
