### 2022-10-31 jjmccollum/teiphy#52 @rbturnbull
Hi @jjmccollum - the revisions look great. I think it might be good to list the other export formats in the summary. I think it's a key feature that it can export to PHYLIP, FASTA etc.
