### 2022-10-31 7a11524
:abc: Added some material on output formats to paper (#52)
