+++ docs/references.bib
@@ -35,7 +35,7 @@
 }
 
 @article{mscswhl_iqtree_2020,
-    author = {Minh, Bui Quang and Schmidt, Heiko A and Chernomor, Olga and Schrempf, Dominik and Woodhams, Michael D and von Haeseler, Arndt and Lanfear, Robert},
+    author = {Minh, Bui Quang and Schmidt, Heiko A. and Chernomor, Olga and Schrempf, Dominik and Woodhams, Michael D. and {von Haeseler}, Arndt and Lanfear, Robert},
     title = {{IQ-TREE 2: New Models and Efficient Methods for Phylogenetic Inference in the Genomic Era}},
     journal = {Molecular Biology and Evolution},
     volume = {37},
@@ -72,7 +72,7 @@
 }
 
 @article{bouckaert_beast_2019,
-    author = {Bouckaert, Remco and Vaughan, Timothy G. and Barido-Sottani, Jo\"{e}lle and Duch\^{e}ne, Sebasti\'{a}n and Fourment, Mathieu and Gavryushkina, Alexandra and Heled, Joseph and Jones, Graham and K\"{u}hnert, Denise and De Maio, Nicola and Matschiner, Michael and Mendes, F\'{a}bio K. and M\"{u}ller, Nicola F. and Ogilvie, Huw A. and du Plessis, Louis and Popinga, Alex and Rambaut, Andrew and Rasmussen, David and Siveroni, Igor and Suchard, Marc A. and Wu, Chieh-Hsi and Xie, Dong and Zhang, Chi and Stadler, Tanja and Drummond, Alexei J.},
+    author = {Bouckaert, Remco and Vaughan, Timothy G. and Barido-Sottani, Jo\"{e}lle and Duch\^{e}ne, Sebasti\'{a}n and Fourment, Mathieu and Gavryushkina, Alexandra and Heled, Joseph and Jones, Graham and K\"{u}hnert, Denise and De Maio, Nicola and Matschiner, Michael and Mendes, F\'{a}bio K. and M\"{u}ller, Nicola F. and Ogilvie, Huw A. and {du Plessis}, Louis and Popinga, Alex and Rambaut, Andrew and Rasmussen, David and Siveroni, Igor and Suchard, Marc A. and Wu, Chieh-Hsi and Xie, Dong and Zhang, Chi and Stadler, Tanja and Drummond, Alexei J.},
     title = {{BEAST 2.5: An advanced software platform for Bayesian evolutionary analysis}},
     journal = {{PLOS} Computational Biology},
     volume = {15},
@@ -84,7 +84,7 @@
 }
 
 @article{pedregosa_scikit-learn_2011,
-    author = {Fabian Pedregosa and Ga{{\"e}}l Varoquaux and Alexandre Gramfort and Vincent Michel and Bertrand Thirion and Olivier Grisel and Mathieu Blondel and Peter Prettenhofer and Ron Weiss and Vincent Dubourg and Jake Vanderplas and Alexandre Passos and David Cournapeau and Matthieu Brucher and Matthieu Perrot and {{\'E}}douard Duchesnay},
+    author = {Fabian Pedregosa and Ga\"{e}l Varoquaux and Alexandre Gramfort and Vincent Michel and Bertrand Thirion and Olivier Grisel and Mathieu Blondel and Peter Prettenhofer and Ron Weiss and Vincent Dubourg and Jake Vanderplas and Alexandre Passos and David Cournapeau and Matthieu Brucher and Matthieu Perrot and \'{E}douard Duchesnay},
     title = {{Scikit-learn: Machine Learning in Python}},
     journal = {Journal of Machine Learning Research},
     volume = {12},
@@ -308,13 +308,33 @@
 
 
[…]
