### 2022-10-31 jjmccollum/teiphy#52 @jjmccollum
@rbturnbull Thanks! I've made that update, and you should be able to see it in the paper now. I'll close this issue and have editorialbot validate the paper again.
