### 2022-10-31 #4879 @jjmccollum
@SimonGreenhill We've made your suggested revisions to the paper and the code, and both should now be accessible in the GitHub repo. I've also had editorialbot regenerate the paper PDF here for easy access. If you have any further comments, they are appreciated!
