+++ pyproject.toml
@@ -35,7 +35,7 @@ python-slugify = "^6.1.2"
 Sphinx = "^5.1.1"
 pytest = "^7.2.0"
 coverage = "^6.4.4"
-nbsphinx = "^0.8.7"
+nbsphinx = "^0.8.9"
 sphinx-rtd-theme = "^1.0.0"
 sphinx-autobuild = "^2021.3.14"
 pre-commit = "^2.15.0"
