### 2022-11-29 #4879 @tresoldi
# Evaluation

The paper presents a high quality software that, as described by the authors, fills a very important void in phylogenetic studies for textual traditions. I can see their work becoming part of a "best practice pipeline" for such analyses, and I will surely start to recommend it.

Not addressing the topics already pointed by @SimonGreenhill :

## Comments on the software

- [ ] While I understand the decision from a philological perspective and for the symmetry with other options (such as for ambiguous states), when considering the most expected usages for the output it would make sense to have the NEXUS format defaulting to "states present".
- [ ] The authors briefly mention how some phylogenetic software (like IQTree) can automatically perform ascertainment correction (usually with the Lewis MK model), but an `--ascertainment` flag would be very welcome for the NEXUS output format in general (or even for all formats). Not only some software expects the user to perform the correction themselves by adding the states to the alignment, but in some cases we might prefer to have it in the model even when the software offers it as an option.

## Comments on the paper

- [ ] the format for Carlson's STEMMA is described as "unique" (lines 13 and 74), which I assume means "exclusive to". The adjective seems superfluous, not only because the format is not "unique" but also because the audience of the paper will most likely know about the format.
- [ ] There is no dispute that critical texts are fundamental to classical phylology (line 35), but the authors use "digital humanities" in a stricter sense than the one adopted by most people, and more as "the on-going digitalization of philological work". This is a common usage in the field of which I am also guilty, but as it has been pointed to me in the past it might be worth talking about "digital turn in philology" or something alike, reserving DH for the sense of "intersection of computing and humanities in general"
- [ ] In line 53, it is possible to say that *no* currently available phylogenetic software expects input in TEI XML, stressing the value of this contribution
- [ ] In line 54, it would be worth mentioning that the format was conceived not only for versatility, but also for promiting digital adoption in stemmatology
- [ ] In line 64, PAUP* is not cited
- [ ] In line 85, from experience there are cases where researchers might prefer or need a long table format, transposing the table (it is not practical for human inspection in most cases, but I would not dismiss it)
- [ ] Installations instructions are given in the repository README file and in the documentation, but not in the draft. It is not an issue, given that it is documented and the procedure is easy and intuitive, but explict instructions are missing.

And, once more, great work @jjmccollum !
