### 2022-11-29 #4879 @jjmccollum
@tresoldi Thank you for these comments! These changes seem straightforward and shouldn't take too long to implement (though I might need @rbturnbull's help on the `--ascertainment` option for NEXUS outputs). I'll open a branch or two for revisions to the paper and additions to the code.

Regarding the citation of PAUP*, I wasn't able to find a paper describing it, but it looks like the website just suggests citing the software as follows:

> Swofford, D. L. 2003. PAUP*. Phylogenetic Analysis Using Parsimony (*and Other Methods).
> Version 4. Sinauer Associates, Sunderland, Massachusetts.

If you had a paper in mind that would be more appropriate to cite, just let us know; otherwise, I will cite the software above.
