### 2022-11-30 jjmccollum/teiphy#55 @jjmccollum
Make `--states-present` default option for NEXUS output
Per the latest comments on https://github.com/openjournals/joss-reviews/issues/4879, the most common expected use case for NEXUS outputs is the use of state symbols as opposed to state frequency vectors. For this reason, NEXUS outputs should be generated with the setting `StatesFormat=StatesPresent` by default. Instead of a `--states-present` command-line option, we should offer a `--frequency` option that sets `StatesFormat=Frequency` instead.

These changes should be made on the `code-revision` branch.
