### 2022-11-30 jjmccollum/teiphy#56 @jjmccollum
Add support for `--ascertainment` option
Per the latest comments on https://github.com/openjournals/joss-reviews/issues/4879:
> The authors briefly mention how some phylogenetic software (like IQTree) can automatically perform ascertainment correction (usually with the Lewis MK model), but an `--ascertainment` flag would be very welcome for the NEXUS output format in general (or even for all formats). Not only some software expects the user to perform the correction themselves by adding the states to the alignment, but in some cases we might prefer to have it in the model even when the software offers it as an option.

I suspect that in NEXUS format, options like this would be specified in the ASSUMPTIONS block. If this specification is dependent on the target software (like IQ-TREE, MrBayes, or BEAST), the separate options may have to be developed for each software. I haven't yet found examples of NEXUS files that include these options.

These changes should be made on the `code-revision` branch.
