### 2022-11-30 jjmccollum/teiphy#57 @jjmccollum
Incorporate latest feedback on paper
Per the latest comments at https://github.com/openjournals/joss-reviews/issues/4879:
- [x] the format for Carlson's STEMMA is described as "unique" (lines 13 and 74), which I assume means "exclusive to". The adjective seems superfluous, not only because the format is not "unique" but also because the audience of the paper will most likely know about the format.
- [x] There is no dispute that critical texts are fundamental to classical phylology (line 35), but the authors use "digital humanities" in a stricter sense than the one adopted by most people, and more as "the on-going digitalization of philological work". This is a common usage in the field of which I am also guilty, but as it has been pointed to me in the past it might be worth talking about "digital turn in philology" or something alike, reserving DH for the sense of "intersection of computing and humanities in general"
- [x] In line 53, it is possible to say that no currently available phylogenetic software expects input in TEI XML, stressing the value of this contribution
- [x] In line 54, it would be worth mentioning that the format was conceived not only for versatility, but also for promiting digital adoption in stemmatology
- [x] In line 64, PAUP* is not cited
- [x] In line 85, from experience there are cases where researchers might prefer or need a long table format, transposing the table (it is not practical for human inspection in most cases, but I would not dismiss it)
- [x] Installations instructions are given in the repository README file and in the documentation, but not in the draft. It is not an issue, given that it is documented and the procedure is easy and intuitive, but explict instructions are missing.

In addition, we have the following paper revisions related to the code revisions:
- [x] Update discussion and examples of the `--states-present` option, as we now have a `--frequency` option instead.

These changes should be made on the `paper-revision` branch.
