+++ teiphy/collation.py
@@ -290,45 +290,11 @@ class Collation:
         nexus_symbols = possible_symbols[:nsymbols]
         return nexus_symbols
 
-    # def get_nexus_equates(self, nexus_symbols: List[str]):
-    #     """Returns a list of one-character multiple-reading symbols for the NEXUS Equate block and a dictionary mapping multistate reading index lists to these symbols.
-
-    #     Note that this will ignore any certainty degrees assigned to these states in the collation.
-
-    #     Args:
-    #         nexus_symbols: A list of nexus symbols for the singleton states.
-
-    #     Returns:
-    #         A dictionary mapping tuples of multiple reading indices to their corresponding symbols in the equate block.
-    #     """
-    #     # NOTE: EQUATE symbols are allowed to be case-sensitive, so we can use uppercase characters for these:
-    #     possible_symbols = (
-    #         list(string.ascii_uppercase)
-    #     )
-    #     # First, populate a set of all reading index tuples that we encounter in the collation:
-    #     reading_ind_tuples_set = set()
-    #     # If there are no witnesses, then no symbols are needed at all:
-    #     if len(self.witnesses) == 0:
-    #         return [], {}
-    #     for wit in self.witnesses:
-    #         for rdg_support in self.readings_by_witness[wit.id]:
-    #             rdg_inds = [
-    #                 i for i, w in enumerate(rdg_support) if w > 0
-    #             ]  # the index list consists of the indices of all readings with any degree of certainty assigned to them
-    #             if len(rdg_inds) > 1:
-    #                 rdg_ind_tuple = tuple(rdg_inds)
-    #                 reading_ind_tuples_set.add(rdg_ind_tuple)
-    #     # Sort the reading index tuples in lexicographical order and map them to the remaining symbols:
-    #     reading_ind_tuples = sorted(list(reading_ind_tuples_set))
-    #     nexus_equates = possible_symbols[: len(reading_ind_tuples)]
-    #     nexus_equate_mapping = {t: possible_symbols[i] for i, t in enumerate(reading_ind_tuples)}
-    #     return nexus_equates, nexus_equate_mapping
-
     def to_nexus(
         self,
         file_addr: Union[Path, str],
         char_state_labels: bool = True,
-        states_present: bool = False,
+        frequency: bool = False,
         ambiguous_as_missing: bool = False,
         calibrate_dates: bool = False,
         mrbayes: bool = False,
@@ -338,13 +304,13 @@ class Collation:
         Args:
             file_addr: A string representing the path to an output NEXUS file; the file type should be .nex, .nexus, or .nxs.
             char_state_labels: An optional flag indicating whether or not to include the CharStateLabels block.
-            states_present: An optional flag indicating whether to use the StatesFormat=StatesPresent setting
-                instead of the StatesFormat=Frequency setting
-                (and thus represent all states with single symbols rather than frequency vectors).
-                Note that this setting will ignore any certainty degrees assigned to multiple ambiguous states in the collation.
+            frequency: An optional flag indicating whether to use the StatesFormat=Frequency setting
+                instead of the StatesFormat=StatesPresent setting
+                (and thus represent all states with frequency vectors rather than symbols).
+                Note that this setting is necessary to make use of certainty degrees assigned to multiple ambiguous states in the collation.
             ambiguous_as_missing: An optional flag indicating whether to treat all ambiguous states as missing data.
                 If this flag is set, then only base symbols will be generated for the NEXUS file.
-                It is only applied if the states_present option is True.
+                It is only applied if the frequency option is False.
             calibrate_dates: An optional flag indicating whether to add an Assumptions block that specifies date distributions for witnesses.
                 This option is intended for inputs to BEAST2.
             mrbayes: An optional flag indicating whether to add a MrBayes block that specifies model settings and age calibrations for witnesses.
[…]
