+++ .github/workflows/iqtree.yml
@@ -20,7 +20,7 @@ jobs:
     - run: poetry install
     - name: Testing
       run: |
-        poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors --states-present example/ubs_ephesians.xml iqtree_example.nexus
+        poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors example/ubs_ephesians.xml iqtree_example.nexus
     - name: Install Phylogenetics Package
       run: sudo apt-get install -y iqtree
     - name: Phylogenetics Run
