+++ docs/references.bib
@@ -14,6 +14,14 @@
     eprint = {https://academic.oup.com/sysbio/article-pdf/46/4/590/19502018/46-4-590.pdf},
 }
 
+@misc{swofford_paup_2003,
+    author = {Swofford, D. L.},
+    title = {PAUP*: Phylogenetic Analysis Using Parsimony (*and Other Methods). Version 4},
+    address = {Sunderland, {MA}},
+    publisher = {Sinauer Associates},
+    year = {2003},
+}
+
 @book{farris_hennig86_1988,
     author = {Farris, James S.},
     title = {{Hennig86, ver. 1.5. Program and Documentation}},
