### 2022-11-30 jjmccollum/teiphy#56 @jjmccollum
For IQ-TREE, the desired behavior is simple; if `--ascertainment` is not specified, then the command-line options for `iqtree` include `-m MK`; if it is specified, the options should include `-m MK+ASC`. The trick is handling this in the NEXUS output generated by `teiphy`. Is there a way to specify this in a block in the NEXUS file so that `iqtree` will read the options from that file?

For MrBayes, we conveniently have a MRBAYES block in the NEXUS file, so this approach to encoding the options in the input file is possible. Specifically, if `--ascertainment` is specified, then `lset coding=variable;` should be included in the MRBAYES block.

I'm hoping that there's a similar way to do this for BEAST2 using the ASSUMPTIONS block of the NEXUS file, since it already anticipates clock model calibration settings in the ASSUMPTIONS block.
