### 2022-11-30 #4879 @tresoldi
@jjmccollum for ascertainment, only adding a uniform character with zeros should be enough. I have seen a small group of people working in computational linguistics experimenting with having a missing value (`?`) in the ascertainment character when the value is missing (so that the ascertainment character value has both zeros and questions marks), but it is not the common practice and would over-complicate things. Maybe for a future version, if somebody asks for it.

As for PAUP* that is the reference I see from time to time, that is perfect.
