### 2022-11-30 #4879 @jjmccollum
@tresoldi Oh, that makes things much easier! Thanks for clarifying that. I should be able to add that feature for most or all output formats (although I don't know if every phylogenetic program we test in our workflows will use the ascertainment character).
