### 2022-11-30 #4879 @rbturnbull
Hi @jjmccollum - there are helpful explanations about correcting for ascertainment bias with a dummy character in the Lewis 2001 paper and in this paper:
Felsenstein, J. (1992), PHYLOGENIES FROM RESTRICTION SITES: A MAXIMUM-LIKELIHOOD APPROACH. Evolution, 46: 159-173. https://doi.org/10.1111/j.1558-5646.1992.tb01991.x

I'm happy to work on this.
