### 2022-11-30 #4879 @rbturnbull
Hi @tresoldi - I've been reading papers on correcting for ascertainment bias and looking through implementations in phylogenetic software programs. I now think that it is actually out of scope for the teiphy tool because there are different ways of doing it that are bound up with the phylogenetic method that you are using. I think it's best for teiphy to just convert the TEI XML data and represent it in nexus, CSV etc and to leave it up to the user to correct for ascertainment bias in the phylogenetic software package that they use. Are you happy with that?
