### 2022-11-30 #4879 @tresoldi
Hi @rbturnbull . The short answer is that yes, I am happy with it.

Longer answer. I was considering the issue this afternoon, and you are right that different models and tools (besides research decisions) would require different approaches for ascertainment bias (even more considering how correction has been used mostly for better dating in biology/linguistics, which works differently for textual traditions).

Having an option for the most common correction (a uniform character with absence) still sounds like a good addition, but not one upon which the paper acceptance would depend. I would be happier with a comment (in the documentation)  explicitly saying that teiphy makes no such correction which left to the user/software, but this is not "mandatory".

As you mention, users can easily perform it from other formats, especially tabular ones. The output for such task could, however, be improved: it might be my fault, but `teiphy` seems to only emit tabular files with probabilities (e.g., 0.333 for the three values of B10K1V1U24-26 for manuscript `01`), given that the `--states-present` option only applies to nexus output.

While parsing the latter is easy, I would like to ask you to investigate if, for users' convenience, it is possible to have a tabular output similar to the matrix of the nexus one when `--states-present` is set. It would be even better if you could provide the results in long table format, something like (optionally using the nexus character value indexes, along with question marks, instead of the actual reading):

```
Taxon,Character,Value
UBS,B10K1V1U24-26,εν εφεσω
P46,B10K1V1U24-26,om.
01,B10K1V1U24-26,
02,B10K1V1U24-26,εν εφεσω
03,B10K1V1U24-26,
04,B10K1V1U24-26,om.
UBS,B10K1V6U22-26,εν τω ηγαπημενω
P46,B10K1V6U22-26,εν τω ηγαπημενω
01,B10K1V6U22-26,εν τω ηγαπημενω
02,B10K1V6U22-26,εν τω ηγαπημενω
03,B10K1V6U22-26,εν τω ηγαπημενω
04,B10K1V6U22-26,εν τω ηγαπημενω υιω αυτου
```

Once more, this is not a request but a suggestion, as it goes beyond your initial scope and does not diminish the validity and importance of your submission. I would be happy to open it as an issue in the teiphy repository once the paper is published, where we could discuss in more detail.
