### 2022-12-01 #4879 @jjmccollum
@tresoldi I think I could implement a `--long-table` option for tabular outputs (NumPy, Pandas DataFrame, CSV, TSV) like the one detailed above. To make things easier, I might just include both the reading symbol and the long form of the reading in consecutive columns:

```
Taxon,Character,Reading,Value
UBS,B10K1V1U24-26,1,εν εφεσω
P46,B10K1V1U24-26,2,om.
01,B10K1V1U24-26,2,om.
02,B10K1V1U24-26,1,εν εφεσω
03,B10K1V1U24-26,2,om.
04,B10K1V1U24-26,?,
UBS,B10K1V6U22-26,1,εν τω ηγαπημενω
P46,B10K1V6U22-26,1,εν τω ηγαπημενω
01,B10K1V6U22-26,1,εν τω ηγαπημενω
02,B10K1V6U22-26,1,εν τω ηγαπημενω
03,B10K1V6U22-26,1,εν τω ηγαπημενω
04,B10K1V6U22-26,2,εν τω ηγαπημενω υιω αυτου
```

For this, I would probably have to coerce all ambiguous readings to look like missing data (i.e., `?`), because duplicating rows for the same taxon and character seems more potentially problematic. How does this sound to you?
