+++ docs/references.bib
@@ -354,3 +354,22 @@
     publisher = {Deutsche Bibelgesellschaft},
     year = {2014},
 }
+
+
+@article{lewis_likelihood_2001,
+	title = {A {Likelihood} {Approach} to {Estimating} {Phylogeny} from {Discrete} {Morphological} {Character} {Data}},
+	volume = {50},
+	issn = {1063-5157},
+	shorttitle = {A {Likelihood} {Approach}},
+	url = {https://doi.org/10.1080/106351501753462876},
+	doi = {10.1080/106351501753462876},
+	abstract = {Evolutionary biologists have adopted simple likelihood models for purposes of estimating ancestral states and evaluating character independence on specified phylogenies; however, for purposes of estimating phylogenies by using discrete morphological data, maximum parsimony remains the only option. This paper explores the possibility of using standard, well-behaved Markov models for estimating morphological phylogenies (including branch lengths) under the likelihood criterion. An important modification of standard Markov models involves making the likelihood conditional on characters being variable, because constant characters are absent in morphological data sets. Without this modification, branch lengths are often overestimated, resulting in potentially serious biases in tree topology selection. Several new avenues of research are opened by an explicitly model-based approach to phylogenetic analysis of discrete morphological data, including combined-data likelihood analyses (morphology + sequence data), likelihood ratio tests, and Bayesian analyses.},
+	number = {6},
+	urldate = {2020-11-08},
+	journal = {Systematic Biology},
+	author = {Lewis, Paul O.},
+	month = nov,
+	year = {2001},
+	pages = {913--925},
+	file = {Lewis - 2001 - A Likelihood Approach to Estimating Phylogeny from.pdf:/Users/rturnbull/Zotero/storage/UFTBJR82/Lewis - 2001 - A Likelihood Approach to Estimating Phylogeny from.pdf:application/pdf},
+}
