### 2022-12-01 #4879 @rbturnbull
Hi @tresoldi - thanks for your comment. That all sounds great. I've added a section in the docs about ascertainment bias here: https://jjmccollum.github.io/teiphy/advanced.html#ascertainment-bias
