### 2022-12-01 4d8fa00
:zap: Added source code, test coverage, and documentation for --long-table option (and tabular outputs in general) (#58)
