+++ teiphy/collation.py
@@ -756,7 +756,7 @@ class Collation:
                             row_ind += 1
                     else:
                         row_ind += len(rdg_support)
-                # Otherwise, adds its coefficients normally:
+                # Otherwise, add its coefficients normally:
                 else:
                     for i in range(len(rdg_support)):
                         matrix[row_ind, col_ind] = rdg_support[i]
@@ -807,45 +807,117 @@ class Collation:
                     matrix[i, j] = disagreements
         return matrix, witness_labels
 
-    def to_dataframe(self, split_missing: bool = True):
+    def to_long_table(self):
+        """Returns this Collation in the form of a long table with columns for taxa, characters, reading indices, and reading values.
+        Note that this method treats ambiguous readings as missing data.
+
+        Returns:
+            A NumPy array with columns for taxa, characters, reading indices, and reading values, and rows for each combination of these values in the matrix.
+            A list of column label strings.
+        """
+        # Initialize the outputs:
+        column_labels = ["taxon", "character", "state", "value"]
+        long_table_list = []
+        # First, populate a dictionary mapping (variation unit index, reading index) tuples from the readings_by_witness dictionary
+        # to the readings' texts:
+        reading_texts_by_indices = {}
+        substantive_variation_unit_ids_set = set(self.substantive_variation_unit_ids)
+        substantive_variation_unit_reading_tuples_set = set(self.substantive_variation_unit_reading_tuples)
+        vu_ind = 0
+        for vu in self.variation_units:
+            if vu.id not in substantive_variation_unit_ids_set:
+                continue
+            rdg_ind = 0
+            for rdg in vu.readings:
+                key = tuple([vu.id, rdg.id])
+                if key not in substantive_variation_unit_reading_tuples_set:
+                    continue
+                indices = tuple([vu_ind, rdg_ind])
+                reading_texts_by_indices[indices] = rdg.text
+                rdg_ind += 1
+            if rdg_ind > 0:
+                vu_ind += 1
+        # Then populate the output list with the appropriate values:
+        witness_labels = [wit.id for wit in self.witnesses]
+        missing_symbol = '?'
+        for wit_id in witness_labels:
+            for vu_ind, rdg_support in enumerate(self.readings_by_witness[wit_id]):
+                # Get the variation unit label for this unit:
+                vu_label = self.substantive_variation_unit_ids[vu_ind]
+                # Populate a list of nonzero coefficients for this reading support vector:
+                rdg_inds = [k for k, w in enumerate(rdg_support) if w > 0]
+                # If this list does not consist of exactly one reading, then treat it as missing data:
+                if len(rdg_inds) != 1:
+                    long_table_list.append([wit_id, vu_label, missing_symbol, missing_symbol])
+                    continue
+                rdg_ind = rdg_inds[0]
+                rdg_text = reading_texts_by_indices[(vu_ind, rdg_ind)]
+                # Replace empty reading texts with the omission placeholder:
+                if rdg_text == "":
+                    rdg_text = "om."
+                long_table_list.append([wit_id, vu_label, rdg_ind, rdg_text])
+        # Then convert the long table entries list to a NumPy array:
+        long_table = np.array(long_table_list)
+        return long_table, column_labels
+
+    def to_dataframe(self, long_table: bool = False, split_missing: bool = True):
         """Returns this Collation in the form of a Pandas DataFrame array, including the appropriate row and column labels.
 
         Args:
+            long_table: An optional flag indicating whether or not to generate a long table with columns for taxa, characters, reading indices, and reading values.
+            Note that if this option is set, ambiguous readings will be treated as missing data, and the split_missing option will be ignored.
[…]
