### 2022-12-01 jjmccollum/teiphy#56 @jjmccollum
@rbturnbull Now that you've added some discussion of this to the documentation, I'll go ahead and close this issue. If there are any outstanding issues related to this, feel free to reopen the issue.
