### 2022-12-01 #4879 @jjmccollum
@tresoldi Okay, I just finished implementing the `--long-table` option as requested! I'm using four columns instead of three, as I described in my last comment, and because of how I loop through the entries in the collation, the rows are ordered by taxon first and character second, but both of these features can be easily adjusted in an editor like Excel.

If all of this looks satisfactory to you, then I'm happy with passing this back up to @majensen! If the paper proceeds to publication, then I will plan to bump the repository (and PyPI entry) to a new release with these features at the same time.
