### 2022-12-01 #4879 @tresoldi
@majensen I recommend the paper for publication, it is a great addition to the TEI ecosystem and to digital stemmatology in general.
