### 2022-12-03 #4879 @majensen
@tresoldi @SimonGreenhill thanks so much for your thoughtful and detailed reviews. @jjmccollum I will run through the paper with my editorial hat on, and will correct, rebuke, and encourage. After this, I will provide some instructions about creating an open archive prior to publication.
