### 2022-12-06 de9ec3b
:zap: Updated example XML file to include notes on intrinsic and transcriptional judgments
