### 2022-12-19 #4879 @majensen
@jjmccollum the paper reads very well and I enjoyed it. It occurred to me that STEMMA's ability to detect contamination is exactly analogous to recombination. There is a large literature on "reticulate evolution", which is just accounting for the fact that recombination pulls in different histories for parts of the same gene. I suppose that using multiple sources to create a given manuscript is essentially this problem. The application I see in work would be to check out algorithms that can detect and display these mixtures and see if TEIphy's encoding could generate a reticulation in the history of the three contaminated witnesses. I could go on because I find this pretty fascinating, but I won't. Check this ref I found https://www.sciencedirect.com/science/article/pii/S2666166722000715 which is methods-oriented and also contains a number of interesting references.
