### 2022-12-19 #4879 @majensen
HOWEVER, don't do that for this paper. I think we are ready to go.
Can you make an open archive of the codebase using Zenodo, FigShare or similar (see this tutorial https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729). If you would report the DOI of the resulting archive here that would be great.
