### 2022-12-19 #4879 @rbturnbull
Thanks @majensen - this is a big area of research for us. We need to keep exploring ways of modelling contamination/reticulation. Thanks for the link for the article!
