### 2022-12-19 #4879 @rbturnbull
Terrific. It'd be good to release to PyPI as well.
