### 2022-12-19 #4879 @jjmccollum
Agreed. But that should happen automatically once we update the release on GitHub, right?
