+++ README.rst
@@ -3,7 +3,7 @@
 .. image:: https://raw.githubusercontent.com/jjmccollum/teiphy/main/docs/img/teiphy-logo.svg
 
 |license badge| |testing badge| |coverage badge| |docs badge| |black badge| |git3moji badge| 
-|iqtree badge| |raxml badge| |mrbayes badge| |stemma badge| |joss badge|
+|iqtree badge| |raxml badge| |mrbayes badge| |stemma badge| |joss badge| |doi badge|
 
 .. |license badge| image:: https://img.shields.io/badge/license-MIT-blue.svg?style=flat
     :target: https://choosealicense.com/licenses/mit/
@@ -37,6 +37,9 @@
 
 .. |joss badge| image:: https://joss.theoj.org/papers/e0a813f4cdf56e9f6ae5d555ce6ed93b/status.svg
     :target: https://joss.theoj.org/papers/e0a813f4cdf56e9f6ae5d555ce6ed93b
+    
+.. |doi badge| image:: https://zenodo.org/badge/222792228.svg
+    :target: https://zenodo.org/badge/latestdoi/222792228
 
 .. end-badges
 
