### 2022-12-19 #4879 @jjmccollum
@majensen Oh yeah! I've seen median-joining networks before. They've actually been used in textual criticism, as well, for the very reasons you've described (the earliest example I can think of is https://link.springer.com/article/10.1023/B:CHUM.0000009290.14571.59).

I've also connected the latest release of `teiphy` to Zenodo. The assigned DOI for all versions of the software is [10.5281/zenodo.7455638](https://doi.org/10.5281/zenodo.7455638).

@rbturnbull, it looks like the IQTREE workflow is now failing for some reason. I'm not sure why, as it was passing the last time I merged a pull request. Do you think it might be an issue with the version of IQTREE being installed in the workflow?
