+++ .github/workflows/iqtree.yml
@@ -22,7 +22,9 @@ jobs:
       run: |
         poetry run teiphy -t reconstructed -t defective -t orthographic -m overlap -m lac -s"*" -s T --fill-correctors example/ubs_ephesians.xml iqtree_example.nexus
     - name: Install Phylogenetics Package
-      run: sudo apt-get install -y iqtree
+      run: |
+        sudo sed -i 's/azure\.//' /etc/apt/sources.list
+        sudo apt-get install -y iqtree
     - name: Phylogenetics Run
       run: |
         iqtree -s iqtree_example.nexus -m MK+ASC -bb 1000
