### 2022-12-19 #4879 @majensen
Great @jjmccollum - are you willing to name the archive with the exact name of the paper? If so, this is preferred by JOSS.
