### 2022-12-19 #4879 @rbturnbull
@jjmccollum I'll create a separate issue about the iqtree workflow. It's just that there's now iqtree executable in the path after it is installed with apt-get. I don't think it is related to our code.
