### 2022-12-19 0c707bc
:tv: trying iqtree2 executable name (#60)
