### 2022-12-19 #4879 @danielskatz
@majensen - this paper is far too long for a JOSS paper, which isn't worth fixing at this point, but is something we should look for in the future...
