+++ docs/references.bib
@@ -1,6 +1,6 @@
 @article{msm_nexus_1997,
     author = {Maddison, David R. and Swofford, David L. and Maddison, Wayne P.},
-    title = {{NEXUS: An Extensible File Format for Systematic Information}},
+    title = {{NEXUS}: An Extensible File Format for Systematic Information},
     journal = {Systematic Biology},
     volume = {46},
     number = {4},
@@ -24,7 +24,7 @@
 
 @book{farris_hennig86_1988,
     author = {Farris, James S.},
-    title = {{Hennig86, ver. 1.5. Program and Documentation}},
+    title = {Hennig86, ver. 1.5. Program and Documentation},
     address = {Port Jefferson Station, {NY}},
     publisher = {James S. Farris},
     year = {1988},
@@ -32,7 +32,7 @@
 
 @article{gc_tnt_2016,
     author = {Goloboff, Pablo A. and Catalano, Santiago A.},
-    title = {{TNT, Version 1.5, Including a Full Implementation of Phylogenetic Morphometrics}},
+    title = {{TNT}, Version 1.5, Including a Full Implementation of Phylogenetic Morphometrics},
     journal = {Cladistics},
     volume = {32},
     number = {3},
@@ -44,7 +44,7 @@
 
 @article{mscswhl_iqtree_2020,
     author = {Minh, Bui Quang and Schmidt, Heiko A. and Chernomor, Olga and Schrempf, Dominik and Woodhams, Michael D. and {von Haeseler}, Arndt and Lanfear, Robert},
-    title = {{IQ-TREE 2: New Models and Efficient Methods for Phylogenetic Inference in the Genomic Era}},
+    title = {{IQ-TREE} 2: New Models and Efficient Methods for Phylogenetic Inference in the Genomic Era},
     journal = {Molecular Biology and Evolution},
     volume = {37},
     number = {5},
@@ -58,7 +58,7 @@
 
 @article{stamatakis_raxml_2014,
     author = {Stamatakis, Alexandros},
-    title = {{RAxML Version 8: A Tool for Phylogenetic Analysis and Post-analysis of Large Phylogenies}},
+    title = {{RAxML} Version 8: A Tool for Phylogenetic Analysis and Post-analysis of Large Phylogenies},
     journal = {Bioinformatics},
     volume = {30},
     number = {9},
@@ -69,7 +69,7 @@
 
 @article{rtmadhllsh_mrbayes_2012,
     author = {Ronquist, Fredrik and Teslenko, Maxim and {van der Mark}, Paul and Ayres, Daniel L. and Darling, Aaron and H\"{o}hna, Sebastian and Larget, Bret and Liu, Liang and Suchard, Marc A. and Huelsenbeck, John P.},
-    title = {{MRBAYES 3.2: Efficient Bayesian phylogenetic inference and model selection across a large model space}},
+    title = {{MRBAYES} 3.2: Efficient {B}ayesian phylogenetic inference and model selection across a large model space},
     journal = {Systematic Biology},
     volume = {61},
     number = {3},
@@ -81,7 +81,7 @@
 
 @article{bouckaert_beast_2019,
[…]
