### 2022-12-19 #4879 @jjmccollum
@danielskatz Thank you for the edits and recommendations! The small changes look good to me, and most of the case changes in the bib file look correct. (Apologies for messing around with the capitalization; I'm used to doing this with the standard LaTeX citation style to achieve American-style capitalization for things like titles, but I didn't think about whether this was the preference for _JOSS_.) I would only suggest the following additional changes:
- In the `bhbr_phylogeny_1998` entry, the book title _The Canterbury Tales_, which occurs within the article title, should be in italics. If this can be achieved by changing that part of the article title to `\emph{{T}he {C}anterbury {T}ales}`, then that would be ideal.
- In the entries `baldwin_fa_2010`, `swh_vorlage_2002`, `swh_pathways_2004`, and `ubs5`, the word "Greek" in the `title` field should be capitalized.
- In the `swh_vorlage_2002` entry, the words "Greek" and "Vorlage" (the latter being a German noun) in the `title` field should be capitalized.
- In the `carlson_text_2015` entry, the word "Galatians" in the `title` field should be capitalized (as it is a proper noun), and the word "Untersuchungen" in the `series` field should be capitalized (per the conventions of German capitalization).
- In the `edmondson_analysis_2019` entry, "Coherence-Based Genealogical Method" is a proper noun, and as such, it should be capitalized.

I have copied these comments into my review of your proposed changes, as well. If you have any further questions, feel free to let me know. Thanks again!
