### 2022-12-19 6f52a72
further bib case changes
