+++ docs/references.bib
@@ -137,7 +137,7 @@
 
 @article{baldwin_fa_2010,
     author = {Baldwin, Clinton S.},
-    title = {Factor Analysis: A New Method for Classifying {N}ew {T}estament Greek Manuscripts},
+    title = {Factor Analysis: A New Method for Classifying {N}ew {T}estament {G}reek Manuscripts},
     journal = {Andrews University Seminary Studies},
     volume = {48},
     number = {1},
@@ -228,7 +228,7 @@
 
 @article{bhbr_phylogeny_1998,
     author = {Barbrook, Adrian C. and Howe, Christopher J. and Blake, Norman and Robinson, Peter},
-    title = {The Phylogeny of The {C}anterbury {T}ales},
+    title = {The Phylogeny of \emph{{T}he {C}anterbury {T}ales}},
     journal = {Nature},
     volume = {394},
     year = {1998},
@@ -249,7 +249,7 @@
 
 
 @article{swh_vorlage_2002,
-    title = {The Greek Vorlage of the {S}yra {H}arclensis: A Comparative Study on Method in Exploring Textual Genealogy},
+    title = {The {G}reek {V}orlage of the {S}yra {H}arclensis: A Comparative Study on Method in Exploring Textual Genealogy},
     author = {Spencer, Matthew and Wachtel, Klaus and Howe, Christopher J.},
     journal = {{TC}: A Journal of Biblical Textual Criticism},
     volume = {7},
@@ -260,7 +260,7 @@
 
 @article{swh_pathways_2004,
     author = {Spencer, Matthew and Wachtel, Klaus and Howe, Christopher J.},
-    title = {Representing Multiple Pathways of Textual Flow in the Greek Manuscripts of the {L}etter of {J}ames Using Reduced Median Networks},
+    title = {Representing Multiple Pathways of Textual Flow in the {G}reek Manuscripts of the {L}etter of {J}ames Using Reduced Median Networks},
     journal = {Computers and the Humanities},
     volume = {38},
     year = {2004},
@@ -271,8 +271,8 @@
 
 @book{carlson_text_2015,
     address = {Tübingen},
-    series = {Wissenschaftliche Untersuchungen zum {N}euen {T}estament},
-    title = {{The Text of Galatians and Its History}},
+    series = {Wissenschaftliche {U}ntersuchungen zum {N}euen {T}estament},
+    title = {{The Text of {G}alatians and Its History}},
     isbn = {978-3-16-153323-5},
     number = {385},
     publisher = {Mohr Siebeck},
@@ -348,7 +348,7 @@
 
 @book{ubs5,
     editor = {Aland, Barbara and Aland, Kurt and Karavidopoulos, Johannes and Martini, Carlo M. and Metzger, Bruce M.},
-    title = {The Greek {N}ew {T}estament},
+    title = {The {G}reek {N}ew {T}estament},
     edition = {5},
     address = {Stuttgart},
     publisher = {Deutsche Bibelgesellschaft},
