### 2022-12-19 8c7f874
:abc: Fixed error in bib file
