+++ docs/references.bib
@@ -186,7 +186,7 @@
 
 @misc{tei_critical_apparatus,
     author = {{TEI Consortium}},
-    title = {TEI P}5: Guidelines for Electronic Text Encoding and Interchange: Critical Apparatus [v.4.4.0]},
+    title = {{TEI P}5: Guidelines for Electronic Text Encoding and Interchange: Critical Apparatus [v.4.4.0]},
     year = "2022",
     howpublished = "\url{https://www.tei-c.org/release/doc/tei-p5-doc/en/html/TC.html}",
     note = "[Online; accessed 3-September-2022]"
