### 2022-12-19 #4879 @danielskatz
@jjmccollum - once this command finishes, can you check the generated pdf and let me know if you are ok with it?  I will do the same, then hopefully we can finish this publication.
