### 2022-12-19 #4879 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
