### 2022-12-19 #4879 @danielskatz
I see line 275 of the bib should be
`    title = {The Text of {G}alatians and Its History},` - I left some extra {}s
