### 2022-12-19 9cc3ad0
:abc: Fixed error in bib file
