+++ docs/references.bib
@@ -272,7 +272,7 @@
 @book{carlson_text_2015,
     address = {Tübingen},
     series = {Wissenschaftliche {U}ntersuchungen zum {N}euen {T}estament},
-    title = {{The Text of {G}alatians and Its History}},
+    title = {The Text of {G}alatians and Its History},
     isbn = {978-3-16-153323-5},
     number = {385},
     publisher = {Mohr Siebeck},
