### 2022-12-19 #4879 @danielskatz
> @danielskatz thanks, I look forward to your guidance on that point. I appreciate your moving forward with this recommendation.

I think it's just to check the paper's length when our process starts, and when it is over 1000 words, think about if parts of the paper don't match the typical JOSS paper (e.g., they include material that should be in the documentation, they include results that would better fit a science paper) and then ask the author to reduce it at that point, before the review starts
