### 2022-12-19 #4879 @danielskatz
It works fine for me - it's likely a DNS caching issue on your side.  (I've learned to wait about 5 minutes after the message "Wait a couple of minutes, then verify that the paper DOI resolves https://doi.org/10.21105/joss.04879" to avoid such issues...  If you are still having problems after a while (~6 hours), please let us know.
